Update utoipa annotations and deprecate unused endpoints
Adds descriptions to many of the utoipa annotations that weren't clear.
Corrects the URLs and arguments for several annotations. Also deprecates
the POST /workspace, commits/index, and workspaces/{workspace_id}/files/delete endpoints, which were unused
