+++ oxen-rust/src/server/src/controllers/data_frames.rs
@@ -26,7 +26,7 @@ use uuid::Uuid;
     get,
     path = "/api/repos/{namespace}/{repo_name}/data_frames/{resource}",
     operation_id = "get_data_frame_slice",
-    tag = "DataFrames",
+    tag = "Data Frames",
     security( ("api_key" = []) ),
     params(
         ("namespace" = String, Path, description = "Namespace of the repository", example = "ox"),
@@ -128,7 +128,7 @@ pub async fn get(
     post,
     path = "/api/repos/{namespace}/{repo_name}/data_frames/{resource}/index",
     operation_id = "start_data_frame_indexing",
-    tag = "DataFrames",
+    tag = "Data Frames",
     security( ("api_key" = []) ),
     params(
         ("namespace" = String, Path, description = "Namespace of the repository", example = "ox"),
@@ -182,7 +182,7 @@ pub async fn index(req: HttpRequest) -> actix_web::Result<HttpResponse, OxenHttp
     post,
     path = "/api/repos/{namespace}/{repo_name}/data_frames/from_directory/{resource}",
     operation_id = "create_data_frame_from_directory",
-    tag = "DataFrames",
+    tag = "Data Frames",
     security( ("api_key" = []) ),
     params(
         ("namespace" = String, Path, description = "Namespace of the repository", example = "ox"),
