+++ oxen-rust/src/server/src/controllers/repositories.rs
@@ -126,6 +126,7 @@ pub async fn show(req: HttpRequest) -> actix_web::Result<HttpResponse, OxenHttpE
     get,
     path = "/api/repos/{namespace}/{repo_name}/stats",
     operation_id = "get_repository_stats",
+    description = "Gets the total number of files, the total size of the files, and the number of different file types",
     tag = "Repositories",
     security( ("api_key" = []) ),
     params(
@@ -243,7 +244,7 @@ pub async fn get_size(req: HttpRequest) -> actix_web::Result<HttpResponse, OxenH
     security( ("api_key" = []) ),
     request_body(
         content = RepoNew,
-        description = "Repository creation payload (JSON or Multipart)",
+        description = "Repository creation JSON",
         content_type = "application/json",
         example = json!({
             "namespace": "ox",
