+++ oxen-rust/src/server/src/services/compare.rs
@@ -5,9 +5,10 @@ use crate::controllers;
 
 pub fn compare() -> Scope {
     web::scope("/compare")
+        // TODO: Deprecate. This belongs in the commits router
         .route(
             "/commits/{base_head:.*}",
-            web::get().to(controllers::diff::commits),
+            web::get().to(controllers::commits::list_commits),
         )
         .route(
             "/dir_tree/{base_head:.*}",
